+++ src/list/whole/view.rs
@@ -490,7 +490,10 @@ where
     fn varied_width(ctx: &Context<Self>, widths: &ColWidths) -> Option<u32> {
         match widths {
             ColWidths::Pixel(widths) => Some(
-                ctx.props().display_info.width_full - widths.iter().filter_map(|x| *x).sum::<u32>(),
+                ctx.props()
+                    .display_info
+                    .width_full
+                    .saturating_sub(widths.iter().filter_map(|x| *x).sum::<u32>()),
             ),
             ColWidths::Ratio(_) => None,
         }